feat(dao): them cac ham truy van san pham ton kho nhieu va sap het hang feat(servlet): lay du lieu ton kho tu DAO va truyen sang view dashboard feat(jsp): hien thi bang ton kho dong va sua loi vo layout Grid Bootstrap    close #113

+++ src/main/java/com/shop/sportstore/dao/OrderDAO.java
@@ -32,8 +32,8 @@ public class OrderDAO extends DBConnection {
         }
         return count;
     }
-    public double calculateMonthlyRevenue(int month, int year) {
-        double total = 0;
+    public long calculateMonthlyRevenue(int month, int year) {
+        long total = 0;
         String sql = "SELECT SUM(TotalPrice) FROM orders WHERE Status = 'COMPLETED' AND MONTH(CreatedAt) = ? AND YEAR(CreatedAt) = ?";
         try (Connection conn = getConnection();
              PreparedStatement ps = conn.prepareStatement(sql)) {
@@ -43,7 +43,7 @@ public class OrderDAO extends DBConnection {
 
             try (ResultSet rs = ps.executeQuery()) {
                 if (rs.next()) {
-                    total = rs.getDouble(1);
+                    total = (long) rs.getDouble(1);
                 }
             }
         } catch (Exception e) {
